move commitTaskPosition to where it should be
